power: Don't set the AP_IDLE flag if a power failure occurred
The AP_IDLE flag prevents the AP from automatically starting up. This is
normally applied when the device has shut down cleanly. If the device
has experienced power loss and has not shut down cleanly, the AP power
state should be restored when power is restored.

With some power supplies, the voltage drops slowly when the AC power is
disconnected. This can cause the EC to detect that there is a power
failure on some higher voltage rails while it continues to be powered
for a short while from a lower voltage rail. In such a scenario, the EC
can actually have enough time to complete all of the S0->S3->S5 power
state transitions and trigger the chipset_shutdown hook that sets the
AP_IDLE flag.

BUG=b:239979405
TEST=Manually tested hotplug behavior on Atlas
BRANCH=none

Change-Id: I82ec3c481fc62c79b5798b4d52f025a2825546f4
Signed-off-by: Pablo Ceballos <pceballos@google.com>