usb_pd_host: notify AP USB MUX status after HARD_RESET event cleared
cros-ec-typec module doesn't fetch the type-c related events in
sequence.  If the cros-ec-typec update the MUX state first and then
update the HARD_RESET events, the previously MUX state will be dropped.

To fix this, when AP notify EC the HARD_RESET event is cleared, and at
EC side, the USB_MUX state is not USB_MUX_NONE, we should notify AP of
the current USB MUX state.

BUG=b:239386916 b:224471170
TEST=1. shutdown AP and reboot AP many times, and ensure the external
     monitor has display
     2. USB_MUX status is updated at cros-ec-typec after HARD_RESET
	event cleared
BRANCH=none

Change-Id: If93341adfc16ea8a5898ee83839d5b62447ffa03
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>