power/mt8186: remove redundant msleep
The original purpose of msleep(500) line is to wait for pmic ready,
it was added in CL:3233783 ps30. After a few code changes, another
power_wait_mask_signals_timeout line was added to provide the same
functionality. Now we are waiting for pmic twice, we should remove the
duplicated msleep call.

BUG=b:237249144
TEST=manually
BRANCH=none

Signed-off-by: Ting Shen <phoenixshen@google.com>
Change-Id: I15e3930cf07c0ce6c225f61727e0566a0644e32a