power: add function power_is_force_shutdown
The discharge of the power is too slow on the brask.
As a result, the EC still has power to execute the code
and which will make the EC set the EC_RESET_FLAG_AP_IDLE
to the system reset flag. With the EC_RESET_FLAG_AP_IDLE
flag being set, the EC will block the device power on
even if the system has an abnormal shutdown. To fix this
problem, we need to add an extra function to query if the
system is shutdown abnormally.

BUG=b:238968132
BRANCH=None
TEST=1. deployed the new ec binary to the device
     2. connect the BJ and suspend the device
     3. remove the BJ and reconnect the BJ to the device
     4. the device boot without pressing the power key

Change-Id: Iad6f1e7163bfe33f0c2806dc33cdf99a4f6b07c4
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>