zephyr: Added test for steelix board specific code
Add a dts file for native_posix tests that want to use npcx gpios and
interrupts. There is more work to be done to make it full featured.
Specifically the npcx pinctrl driver really needs an emulator.

Add test for the board specific code in steelix. This is mostly a sample
on how to do this, more than being that useful of a test on it's own.

BRANCH=None
BUG=b:242032118
TEST=./twister --coverage -T zephyr/test/kingler

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: I9436a447e7cd4f3baa743c77d4b2d89bafeb0c9c