8042: Prefix data_port_state enum values
The 8042 keyboard code actually has two state machines embedded into
one, the AT Keyboard, and the 8042 controller. This change makes it
clear which state belongs to which device. Ideally the two state
machines would be split at a later point.

STATE_NORMAL was renamed to STATE_ATKBD_CMD since the 8042 controller is
directly passing the data byte to the keyboard.

BUG=b:242592141
BRANCH=none
TEST=build guybrush

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: Ib3d8ca36d933a457ea8353d44862e6e27da72a2f