apro: generate the hash with possible factory flags
The AP RO flags may have been non-zero when the factory generated the
hash. The stored hash will not match finalized firmware since it was
generated with non-zero gbb flags and the gbb flags are set to 0 during
finalization.

Cr50 can try to match the saved hash by using factory flags to calculate
the AP RO hash. As long as the GBB flags are actually set to 0 it should
be ok to try calculating the hash with a limited set of possible factory
flags. Try to match the saved hash using GBB flags 0 to calculate the
hash. If that doesn't match, cycle through the rest of the possible
factory flags to see if any of them generate the saved hash. If none of
the factory flags work, fail verification.

This change only has a factory flag value of 0. We'll set possible
factory flags in a followup CL.

BUG=b:236844541
TEST=manual
	# add 0x39 to possible_factory_flags
	# Set GBB flags to 0x39
	/usr/share/vboot/bin/set_gbb_flags.sh 0x39
	# save the hash with GBB 0x39
	ap_ro_hash.py FMAP GBB

	# Verify AP RO verification fails because flags are 0x39
	[349.029624 enable_spi_pinmux: AP]
	[349.030178 tpm_rst_asserted]
	[349.032382 spi_hash_pp_done: AP]
	[349.137962 validate_gbb: invalid flags 39]

	# reboot cr50 to release ec reset
	> reboot

	# Set GBB flags to 0
	/usr/share/vboot/bin/set_gbb_flags.sh 0

	# Verify ap ro verification passes.
	[621.269874 tpm_rst_asserted]
	[621.271846 spi_hash_pp_done: AP]
	[621.346050 validate_gbb: ok]
	[621.346897 usb_spi_sha256_update: 3c0000:1000]
	[621.361793 Using 0 for GBB flags.]
	[621.362625 usb_spi_sha256_update: 3c1000:c]
	[621.369369 usb_spi_sha256_update: 3c1010:2ef0]
	[621.400642 Calculated digest 556504e697ea...]
	[621.401393 Stored digest 65e3781beb43a6a7...]
	[621.402297 usb_spi_sha256_update: 3c0000:1000]
	[621.432907 Using 239 for GBB flags.]
	[621.433780 usb_spi_sha256_update: 3c1000:c]
	[621.436889 AC: -F]
	[621.442520 usb_spi_sha256_update: 3c1010:2ef0]
	[621.473293 matched gbb 239]
	[621.474246 spi_hash_disable]
	[621.475353 AP RO PASS!]
	[621.518756 AP UART off]

Change-Id: I17d191abada342263ea246911ce47ac24dbb940c
Signed-off-by: Mary Ruthven <mruthven@chromium.org>