zephyr: Check return codes in espi
There are a number of espi_{read,write}_lpc_request calls in the shim
espi code that don't check return values.  As it happens the espi
emulator returns errors for all of these calls, and so the returned data
is garbage, and this results in flaky tests.

Add checks for the return code, and don't use any bogus data read.  The
code still has a lot of void functions that could hide problems.

I just guessed on which failures should exit the function and which
should continue on.

BRANCH=None
BUG=b:243023286
TEST=./twister

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: I8db463461511db3391318ae298197ad81e719427