power: Don't set the AP_IDLE flag if HW WP is deasserted
The AP_IDEL flag will be set up when the system shutdown. It will cause
the factory to update the Cr50 FW finish cannot auto boot up. It needs
to press the power button to boot up the system.

Normally HW WP is deasserted in the factory process. So it can avoid
the AP_IDLE flag being set up when the system shutdown after updated
Cr50 FW.

BUG=b:236677296
BRANCH=none
TEST=The system can auto boot up after update the Cr50 FW.

Signed-off-by: Matt Wang <matt_wang@compal.corp-partner.google.com>
Change-Id: Iada4ec6efc430c1b455b358dc4d704fb2d2c6bbd