gitlab: Support running Twister in Gitlab CI
This CL aims to get Twister working in the Gitlab CI using a GCC
toolchain. It also includes some cleanup of the directory structure and
paths in the CI build. The previous approach of checking out zephyr and
all zephyr modules as subdirectories under the EC checkout breaks the
way Twister searches for tests, as all upstream Zephyr tests get
included when we search the EC dir. A detailed layout of the new
directory structure is included as a comment in the YAML file.

Also bumps up the timeout on several driver tests as they were timing
out in CI.

BRANCH=None
BUG=None
TEST=gitlab-runner exec ... {ec_coverage, zephyr_coverage,
zephyr_boards_coverage}

Signed-off-by: Tristan Honscheid <honscheid@google.com>
Change-Id: I80697bd02be363d8462f48abc4b6f0a2394a8926