port80: Add a config to avoid logging port80 codes
Add a config item to disable logging port80 codes as
they are received. This reduces the overhead at AP startup
on the logging. It has been observed on the ITE81302 that
logging the port80 codes causes significant overhead that
may lead to watchdog timeouts.

The port80 code history is still available via the 'port80'
EC console command.

BUG=b:243607643
TEST=zmake build nereid; flash & run
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I543b0391bcf474c9d040baa98ff75f6e2b9fd5a9