rex: Make SOC_PWROK/SYS_PWROK signals TTL instead of OD
Both SOC_PWROK and SYS_PWROK signals are designed to be TTL outputs,
but they were being configured as open drain outputs in their power
signals node.

In addition, this CL adds an internal pullup to SEQ_EC_ALL_SYS_PG as
a work-around for the missing pullup on the board.

Lastly, for the gpio flags field, 0 was replaces with GPIO_ACTIVE_HIGH
for readability.

BRANCH=none
BUG=b:243638916
TEST=zmake build rex

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: Ifd7f88b863d5468fbcfa979d23b0071f429903d1