usbc: Add exiting condition for Unattached.SNK
In USB Type-C spec, for exiting from Unattached.SNK
state(under section-4.5.2.2.3.2), below is one of the
condition to take care of:
A DRP shall transition to Unattached.SRC within
tDRPTransition after the state of both CC pins is
SNK.Open for tDRP − dcSRC.DRP ∙ tDRP, or if directed.
Case Relevance of this CL is when we unplug a self
powered device(TBT), where DRP state in S0 is expected
to be SRC. For ports, where DRP STATE is set as
PD_DRP_FORCE_SOURCE(board specific), a TBT unplug will
not let the port to go to low power mode and role
will continue to be Unattached.SNK, making any further
connection of a bus powered devices like pendrive to go
undetected. This spec rule will take care of exiting from
SNK to SRC and based on the condition, it can further go to
low power mode as well.

BUG=None
BRANCH=None
TEST=Verified TBT/USB4 disconnect and  further Pendrive connection
on ADL/RPL RVP, where DRP state is defaulted as PD_DRP_FORCE_SOURCE.
(dependent CL for this Board specific DRP change-3855157)

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: I57ef5db558d5cd4d33211327ec24fe6b4981546f