TCPMv2: Release TCPC driver on port suspend request
We need to release the TCPC driver from the TCPC when we want to suspend
all operations on the TCPC. In practice, this means we disable ALERTs on
the TCPC so the driver is not invoked.  We introduced this functionality
in TCPMv1 but it never made it to TCPMv2, so add it to TCPMv2.

BUG=b:240458902
BRANCH=none
TEST="pd 2 suspend ; pd 2 resume" no longer stops port 0 due to
	interrupt storm. /sys/class/typec shows devices are enumerated
	on resume.

Change-Id: If6cc99a7a3392fd1ddfdf795074bc9f406b2555e
Signed-off-by: Caveh Jalali <caveh@chromium.org>