test: add tests for crash console command
Adds basic tests for the crash console command involving ASSERT()
crashes. This change introduces a new ASSERT mock which by default
passes the call over to __ASSERT_NO_MSG (which is the default behavior
when setting CONFIG_ASSERT=y. Tests can then override this either in
the test or in the before() function.

BRANCH=none
BUG=b:236074360
TEST=twister -s zephyr/test/drivers/drivers.default

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: I59c24bc4ed4ef81dde5cce16ac0ae4ef96f124cc