rex: Fix the gpio pin type for ec_edp_bl_en
The gpio device tree node for ec_edp_bl_en was being configured as an
open-drain signal, but this should be a TTL with low as its initial
state. This CL corrects this error. The GPIO spreadsheet that's being
used as an input for pinmap needs to be updated. This CL provides a
fix to enable the display panel to work until the tool that creates
the input for pinmap is updated.

BRANCH=none
BUG=b:243919230
TEST=zmake build rex

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I4f8db1e150572df58c1a97d1f509a556383eb0d1