shotzo: remove battery setting and charger task
Based on the shotzo HW design, we don't include a battery even though
the design has the charger ic sm5803. In order to support this HW design,
we remove the battery setting from the board.h, remove battery.c and
CHARGER task from the board customization. To avoid the link error, we
implement some charger APIs in the charger.c which will be called by
sm5803.c and charge_manager.c

BUG=b:242949656
BRANCH=none
TEST=make -b shotzo
    flash the ec.bin to the device and checked the behavior

Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>
Change-Id: I63eeafecb0af0325cda6e72470a9826702611091