rex: Fix the pin type for usb_c0_rt_3p3_sx_en
This GPIO was being set as an open drain output, but it should be a
push-pull. This CL modifies its entry in generated.dts. There is a
TODO to review the EC GPIO input spreadsheet so that generated.dts
will not require edits. This change is being added now to support
power on activies.

BRANCH=none
BUG=b:244441996
TEST=zmake build rex. Verified that HB is powered up when
usb_c0_rt_3p3_sx_en is driven high.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: Ia378a5772a41214c0f49de053d53d5af82c104e8