charger: support the charger without battery
In order to support the HW design with charger but doesn't include
the a battery, we exclude the charge_state_v2.c if we define the
CONFIG_CHARGER but not define the CONFIG_BATTERY. Besides, we also
implement some default functions to avoid the link error.
Note: remove the charge_prevent_power_on from the board.c of
rainier/mchpevb1 and use the default function.

BUG=b:242949656
BRANCH=none
TEST=build pass with
     1. CONFIG_CHARGER + CONFIG_BATTERY
     2. CONFIG_CHARGER + no CONFIG_BATTERY
     3. make buildall

Change-Id: I32c9b03aaf87b641af2b337db1f3bae838464a41
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>