nissa/vif: do not indicate SVDM identity is supported
All Nissa boards are DFP-only, so SVDM identity is unsupported because
no modes (not USB4 nor TBT) are supported. This was previously set to
true because the VIF editor tool forces it on for unclear reasons, but
per the VIF file description there is no reason it can't be disabled.

BUG=b:242671457
TEST=none
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Ia0bd72631634487f0493d9fd3dbebeddb7fbbfdb