zephyr/usbc: implement a "default" SVDM identity
Recent PD specifications require that devices with multiple DFPs respond
to Discover Identity commands, even if they don't support any modes as
responder. This implements a default handler for those commands
appropriate for such devices that reports DFP-only operation, enabled by
config and default-on for devices that aren't DRDs nor need extra SVDM
support (such as for USB4 or TBT).

This is implemented only for Zephyr (no equivalent implementation for
legacy EC-OS) because it is expected that new devices which might
benefit from it will all be using Zephyr.

BUG=b:242671457
TEST=nereid now responds to Discover Identity
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I2d11ced6928ae65ca00d20a1646509281080243b