test: Add test for read_match_secret succeed
Add test for fp_command_read_match_secret when the number of matched
finger is within the allowed range of [0 5[, the state deadline value is
valid, the state is readable with the correct matched finger number, and
a valid positive_match_salt is derived. The test should succeed and
return the message EC_SUCCESS. The response positive_match_secret should
also align with the expected one given a tpm_seed, positive_match_salt,
and user_id.

BRANCH=None
BUG=b:242720387
TEST=make run-fpsensor_state
TEST=make runhosttests

Signed-off-by: Firas Sammoura <fsammoura@google.com>
Change-Id: I21f9a621d5eb6cd8a4635a83e930d74cb45f4d7f