rex: Use mechanical and not GSC power button GPIO
Rex has a GPIO for both the mechanical power button signal and the
power button signal passed from the GSC. The design was to use the
mech power button as the EC wake source and the GSC power button
signal is not required.

This CL moves the required enum power button from the GSC power button
gpio node to the mechanical power button node. In addition, the GSC
power button signal is removed as a PSL wake source.

BUG=b:243353079
BRANCH=none
TEST=zmake build rex

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I4edf3e02ce51deb7d558d67a8dfdbdb20ffb9665