retimer: HB: retry initial read attempt if it fails
For HB, the i2c controller needs more delay to be ready.
CL:3873652 increases the delay after retimer reset, but the 50 msec
fixed delay causes issues with the retimer getting configured
properly.

This patch sets the delay back to 1ms, but adds
an additional read attempt to give sufficent time for the i2c
controller to respond.

BUG=none
BRANCH=none
TEST=build, and no retimer read failure.

Signed-off-by: Li Feng <li1.feng@intel.com>
Change-Id: I4f12c08ae1f16001791242972fe91704c8e8afae