zephyr/console: Return number of bytes consumed by console buffer
Return the number of bytes consumed by the console buffer. This is not the same as the number of bytes written since null bytes are not
written. For now this will always be 0 or len.

This change will allow callers to perform an action when the request
fails, such as printing a warning or retrying.

BUG=b:243709788
BRANCH=None
TEST=Boot skyrim, observe buffered console, pass default.console test.

Change-Id: I97bf5a0776c123abbaed84b5ca7e58e903c0f20d
Signed-off-by: Rob Barnes <robbarnes@google.com>