zmake: Use vpython3 in cmake
Add cmake variable to force it to use vpython3 instead of auto-detecting
the path to python.  This is slightly fragile due to logic in
zephyr/main/cmake/modules/python.cmake. You can pass in PYTHON_PREFER
but it will extract the version from that binary and run
/usr/bin/python${ver}, which is very strange. Pass Python3_EXECUTABLE
to bypass all of that logic, but it has to be exactly the version in
PYTHON_MINIMUM_REQUIRED or it will fail. We'll need to make sure we keep
the vpython version matched with PYTHON_MINIMUM_REQUIRED in the future.

Update the instructions to run ram_report because it works inside the
chroot now.

Added some missing packages to twister_launcher.py as well. I'm not sure
why this change made them start to be required, but they are.

BRANCH=None
BUG=b:243936551
TEST=zmake build -a

Cq-Depend: chromium:3892212
Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: Ied285ce0b69b33a9a3b016c4bd07cab0b6a13b89