ish: add HOSTCMD task check
host_packet_receive() is defined when HOSTCMD task is enabled; so add a
check of HOSTCMD; otherwise there is build error for board which doesn't
enable HOSTCMD task.

One of ISH solutions is to provide power management functionalities. It
powers on ISH and puts it in deep sleep. HOSTCMD task is disabled in this
case to avoid waking up ISH periodically.

BUG=b:234136500
BRANCH=none
TEST=build tglrvp_ish

Signed-off-by: Li Feng <li1.feng@intel.com>
Signed-off-by: Guangfu Hu <guangfu.hu@intel.com>
Change-Id: I64cd6ad4d292847d477f08708ece4d41b4e94650