util/getversion.sh: Fix version when not in a git repo
When building with "FEATURES="test" emerge-hatch chromeos-ec" after using
"cros-workon-hatch chromeos", the working directory is not a git
repository (the files are just copied), so "git describe" in
getversion.sh fails and the logic falls back to creating its own version
number. The test in test/version.c explicitly checks for a version that
begins with "v2" and has a hash length of at least 8, so update the
version string from this case to match.

BRANCH=none
BUG=b:246424843
TEST=FEATURES="test" emerge-hatch chromeos-ec

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: Ie06f9edc2a228eef0162be10e1b1abe123ce6bc7