gitlab: Remove cache and checkout bydate
Remove the cache, since it only works with shared runners, but keep the
needs tags or else Gitlab will run all the build stage jobs before
starting any of the test stage jobs, and that is slower.

Get the timestamp of the latest commit, and use that timestamp to
checkout the corresponding cls in the zephyr and ancillary repos.

Retry git clone calls because they actually fail pretty often.

BRANCH=None
BUG=b:244590155
TEST=Pushed to gitlab coverage branch
https://gitlab.com/zephyr-ec/ec/-/pipelines/640483665

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: Ie66d5c97349ef17eba302387c3cd525dfbaa82ba