typec: Clear stale hard reset event when ready
When entering PE_SNK_Ready or PE_SRC_Ready, clear the
PD_STATUS_EVENT_HARD_RESET flag so that the kernel does not
receive stale events (which can disable the USB port).

The hard reset can be triggered if the EC cannot distinguish
between a reset and a power-on reboot (such as occurs with the IT81302),
and so is flagged as a power-on.
A power-on reset flag will reset the PD connection.
The hard reset PD status event flag is set, and then PD negotiation
may proceed and complete. Eventually the kernel starts, and will read
the PD status, and see the (now stale) hard reset event flag,
and take action to disable the USB port.

BUG=b:231401624
TEST=zmake build nereid; Enter recovery via keyboard
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: Icc179d19e0a0b6fce215021e464516471cdd47cd