i2c-pseudo: Add other documentation formats.
This includes the RST format from patch for latest upstream kernel,
as well as HTML, MD, and PDF generated from the RST.

This is in preparation for changing the kernel<->userspace command
protocol, to be proposed initially via a document diff/patch.

DO NOT SUBMIT this version of the patch.  This patch will change to
replace the text file with markdown, and that's all.

BRANCH=none
BUG=none
TEST=viewed the markdown doc in Gitiles

Change-Id: I332ba9098a1379b2f56efbaab84ed5498ef43d34
Signed-off-by: Matthew Blecker <matthewb@chromium.org>