i8042: Send command response from priority queue
This patch also makes KEYPROTO task hold sending a scancode when it
receives SETLEDS command until it returns ACK to the second byte (and
leaves STATE_ATKBD_SETLEDS).

This patch also removes and repositions some kblog_put calls because
checking OBF and writing to DBBOUT must be done as atomically as
possible to minimize the race condition.

BUG=b:237981131
BRANCH=None
TEST=Taniks. Press search+alt then 'k' 100 times.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I7ccfae99b3657ead5fa9e3c337db623aaffdb0bc