motion_sense: Trigger AP interrupt based on sensor timestamp
Currently, interrupt trigering is based on the time when the motion
sensor task is scheduled. If the task is delayed for processing a
first sample but not a second one, the second sample will be delayed
until the third sample arrive:

/ sensor interrupt       / sensor interrupt     / sensor int
|       / task scheduling| / task scheduling    | / task
+-delay-+----------------+-+--------------------+-+
<--------- 10 ms --------X------------ 10 ms ----->
        <------- < 10 ms ----> event delayed until the
                               next sample.

To address the issue:
1 - add more leeway to schedule the interrupt. Instead of 3us, allow to
  trigger an interrupt if the delta between samples is more than the
  expect delta minus 1ms.
2 - Use the sensor timestamp provided to the FIFO, keep sensor timestamp
for every sensor.

When the timestamp difference is larger than the requested rate, we
schedule an interrupt.

Fixes: 0c71c474869 ("motion sense: Calculate loop time based on sensor needs")

BUG=b:242263033
BRANCH=none
TEST=Ran "tast run localhost:22552 hardware.SensorIioserviceHard"

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I9ccf08c1180820e699cf84069df3e4541768988d