USB Mux: Unify drivers for safe mode
When safe mode was first introduced, it was added only to drivers on the
boards with virtual muxes and the mode wasn't sent to other drivers.

Since safe mode is defined as an input for the mux API, ensure all
drivers can handle this input.  For now, set up all drivers which did
not have handling to handle it the same as None, which was previously
sent to them.

LOW_COVERAGE_REASON=20 of the changed drivers have no emulators
BRANCH=None
BUG=b:234772074
TEST=make -j buildall; search for all files containing a usb_mux_driver
without any reference to USB_PD_MUX_SAFE_MODE and ensure they are
correct to not reference it

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ib927057bbf1638d09000bdad98bc096a6e103611