nissa: Invert the USB-A limit select
The output ilimit select should be inverted so that
the EC output is 1 when current limiting is enabled (SDP), and 0
when current limiting is disabled (CDP).

The default mode is CDP (charging downstream port), and
internally the EC assumes that setting an ilimit of 1 will enable
the higher current, so this value needs to be inverted before
setting the GPIO output.

BUG=b:246504619
TEST=zmake build nivviks; gpioget usb_a0_ilimit_sdp
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: Ieb56ded563f61f0c41d280e7cf65bbc534e6fd50