zephyr/usbc: default-disable USB4
CONFIG_USB_PD_USB4 is default-disabled for EC-OS builds, so do the same
for Zephyr. This makes it much easier to discover which devices support
USB4 by requiring positive confirmation of support, rather than relying
on projects to disable it if not supported which may be less likely for
developers to notice.

Projects that do support USB4 (or may support it but it's difficult to
tell either way) now have it enabled explicitly.

BUG=none
TEST=zmake testall
BRANCH=none

Change-Id: Ib281f15cbb2cfbbaf0c5c3dbb2658405bbfbd423
Signed-off-by: Peter Marheine <pmarheine@chromium.org>