zephyr/usbc: default-disable TBT mode support
CONFIG_USB_PD_TBT_COMPAT_MODE is default-disabled for EC-OS builds, so
do the same for Zephyr. This makes it much easier to discover which
devices support TBT mode by requiring positive confirmation of support,
rather than relying on projects to disable it if not supported which may
be less likely for developers to notice.

Projects that appear to support TBT mode but don't explicitly enable it
are changed to enable it to avoid any undesired configuration change.

BUG=none
TEST=zmake testall
BRANCH=none

Change-Id: Iedc40f5529e6e7f1f9d4605b5364fd76b957934b
Signed-off-by: Peter Marheine <pmarheine@chromium.org>