zephyr: tests: Move non-Zephyr code out of i2c_controller.c
This CL moves 76 SLOC that is not used in Zephyr out of i2c_controller.c
and in to i2c_controller_cros_ec.c. Add a guard to that file so it
cannot be built with CONFIG_ZEPHYR defined.

BRANCH=None
BUG=b:248311206
TEST=make BOARD=brya -j
LOW_COVERAGE_REASON=No new code, moving non-Zephyr code to new file

Signed-off-by: Tristan Honscheid <honscheid@google.com>
Change-Id: I3f71c15baf87d2e59851022d784f2257521cfef6